+++ client/src/components/PetMatchRoutes.js
@@ -2,8 +2,10 @@ import React, { useEffect, useState } from 'react'
 import { Route, Routes } from 'react-router-dom'
 import { useLazyQuery, useMutation } from '@apollo/client';
 import auth from '../utils/auth'
-import { ADD_PET, GET_PROFILE } from '../utils/mutations';
+import { ADD_PET, GET_PROFILE, LOGIN_USER } from '../utils/mutations';
+import { useNavigate } from "react-router-dom";
 
+import Nav from './nav'
 import Home from './home';
 import Profile from './profile';
 import Mission from './mission';
@@ -15,11 +17,24 @@ import SignUp from './signup';
 const PetMatchRoutes = () => {
   const [user, setUser] = useState(auth.getProfile())
   const [pets, setPets] = useState([])
+  const [shouldRedirect, setShouldRedirect] = useState(false)
+
   const [addPet, { petError, petData }] = useMutation(ADD_PET)
+  const [login, { logInerror, LogInUserData }] = useMutation(LOGIN_USER);
+
+  const navigate = useNavigate()
+
   
   const [getPets, { loading, error, data }] = useLazyQuery(GET_PROFILE, {
     variables: { id: user?.data._id  },
   }, [user, pets]);
+
+  useEffect(() => {
+    if (shouldRedirect) {
+      console.log('here')
+      navigate('/profile')
+    }
+  }, [shouldRedirect, navigate])
 
   useEffect(() => { 
     async function fetch() {
@@ -27,7 +42,12 @@ const PetMatchRoutes = () => {
       setPets(pets.data.user.pets)
     }
     if (user) fetch()
-  }, [getPets])
+
+    if (shouldRedirect) {
+      console.log('here')
+      navigate('/profile')
+    }
+  }, [getPets, user, shouldRedirect, navigate])
 
   const likePet = async pet => {
     try {
@@ -52,15 +72,29 @@ const PetMatchRoutes = () => {
     }
   }
 
+  const logInUser = async (formState) => {
+    try {
+      const { data } = await login({
+        variables: { ...formState },
+      });
+      auth.login(data.login.token);
+      setUser(auth.getProfile())
+      setShouldRedirect(true)
+    } catch (e) {
+      console.error(e);
+    }
+  }
+
   return (
     <>
+      <Nav user={user}/>
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/donate" element={<Donate />} />
         <Route path="/ourmission" element={<Mission />} />
         <Route path="/profile" element={<Profile pets={pets}/>} />
         <Route path="/pets" element={<Pet pets={pets} likePet={likePet} />}/>
-        <Route path="/login" element={<Login />} />
+        <Route path="/login" element={<Login logInUser={logInUser} />} />
         <Route path="/signup" element={<SignUp />} />
       </Routes>
     </>
