+++ client/src/components/login.js
@@ -1,26 +1,12 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import '../styles/login.css'
-import { useNavigate } from "react-router-dom";
-
-
-import Footer from "./footer";
-
-import { Link } from 'react-router-dom';
-
-import { useMutation } from '@apollo/client';
-import { LOGIN_USER } from '../utils/mutations';
-
-import Auth from '../utils/auth';
 
 const Login = (props) => {
   const [formState, setFormState] = useState({ email: '', password: '' });
-  const [login, { error, data }] = useMutation(LOGIN_USER);
-  const navigate = useNavigate()
-  const [shouldRedirect, setShouldRedirect] = useState(false)
   // update state based on form input changes
   const handleChange = (event) => {
     const { name, value } = event.target;
-
+    
     setFormState({
       ...formState,
       [name]: value,
@@ -30,29 +16,11 @@ const Login = (props) => {
   // submit form
   const handleFormSubmit = async (event) => {
     event.preventDefault();
-    try {
-      const { data } = await login({
-        variables: { ...formState },
-      });
-      Auth.login(data.login.token);
-      setShouldRedirect(true)
-    } catch (e) {
-      console.error(e);
-    }
-
-    // clear form values
-    setFormState({
-      email: '',
-      password: '',
-    });
+    console.log(props)
+    props.logInUser(formState)
   };
 
-  useEffect(() => {
-    if (shouldRedirect) {
-      console.log('here')
-      navigate('/profile')
-    }
-  }, [shouldRedirect, navigate])
+  
   return (
     <>
       <div className="jumbotron jumbotron-fluid">
